ERR: (double) casts are back to prevent ambiguities on SGI's

+++ Code/BasicFilters/itkExpImageFilter.h
@@ -37,7 +37,7 @@ namespace function {
     ~Exp() {};
     inline TOutput operator()( const TInput & A )
     {
-      return (TOutput)exp(A);
+      return (TOutput)exp((double)A);
     }
   };
